Move commands to another package
