+++ src/main/kotlin/net/axay/blaubot/commands/api/CommandManager.kt
@@ -6,7 +6,7 @@ import dev.kord.core.event.interaction.InteractionCreateEvent
 import dev.kord.core.on
 import kotlinx.coroutines.flow.collect
 import net.axay.blaubot.Manager
-import net.axay.blaubot.commands.*
+import net.axay.blaubot.commands.implementation.*
 import net.axay.blaubot.commands.implementation.anime.AnimeSearch
 import net.axay.blaubot.commands.implementation.anime.AnimeTerms
 import net.axay.blaubot.commands.implementation.anime.RandomAnime
